features/test_runner/requirements.txt
flask==2.3.3
playwright==1.40.0
pandas==2.1.1
openpyxl==3.1.2
APScheduler==3.10.4
python-dotenv==1.0.0
